tiscript v.4.0.1.0 New: 1) Color and Length data types, see doc/Color.htm and doc/Length.htm 2) Stream.scanf() and String.scanf(), type safe version of correspondent C runtime function. 3) tiscript.h (public API) updates. In particular tiscript_class_def::on_gc_copy() - allows to maintain weak reference on the 'this' scripting object in native classes.
Various minor fixes.

+++ trunk/tool/tl_tokenizer.cpp
@@ -985,7 +985,7 @@ DATE_LITERAL:
   {
     int64 li = v.get_int64();
     int64 dollars = li / 10000;
-    uint cents = uint64(li) % 10000;
+    uint cents = uint(uint64(li) % 10000);
     
     i64tow sd(dollars);
     out.put(sd,sd.length());
